+++ test/unit/XmlItemContentTokenizerTest.php
@@ -21,8 +21,6 @@
 
 use oat\tao\test\TaoPhpUnitTestRunner;
 
-include_once dirname(__FILE__) . '/../includes/raw_start.php';
-
 /**
  * This class aims at testing tao_helpers_Xhtml.
  *
