+++ test/unit/pack/encoder/Base64EncoderTest.php
