+++ test/unit/pack/encoder/NoneEncoderTest.php
